Add conditional child profiles, config sanity checks
`processes` is the only property a child profile shares with its parent,
but any properties the child has, the parent must also have in order to
be able to revert the affinity and priorities in effect if the condition
is not true. In this case, it is initially only implemented for when the
fullscreen optimization is applied. Use case - lowering affinity and
priority of demanding processes while a game has focus, in addition to
the increased priority already given when `fullscreenPriority` is `true`
in the app config.