Fix permissions issue when run as own admin user
For some reason UAC privileges aren't enough to modify system processes,
but launching through an elevated powershell instance seems to work.
This runs a VBS wrapper, which runs powershell hidden through an
elevation utility, which runs another VBS wrapper that actually launches
the node app so it can stay hidden. Not very intuitive, but it works for
now.